float instead of int

+++ models.py
@@ -30,8 +30,8 @@ def weighted_crossentropy(weight=None, boundary_weight=None, pool=3):
 
         if boundary_weight is not None:
             y_true_avg = K.pool3d(y_true, pool_size=(pool,)*3, padding='same', pool_mode='avg')
-            boundaries = K.cast(y_true_avg >= K.epsilon(), 'int8') \
-                         * K.cast(y_true_avg <= 1 - K.epsilon(), 'int8')
+            boundaries = K.cast(y_true_avg >= K.epsilon(), 'float32') \
+                         * K.cast(y_true_avg <= 1 - K.epsilon(), 'float32')
             loss *= (boundary_weight * boundaries) + 1
 
         return K.mean(loss)
